Early Design + component / pages routing
Early design and routing between bcomponents and views / pages
